+++ docker/utils/fixtures/non-json.manifest.json
