+++ docker/utils/fixtures/v2s2.manifest.json
